+++ Scripts/Controllers/DSC_Dialogue_ImageGroupController.cs
@@ -1,10 +1,11 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using DSC.UI;
 
 namespace DSC.DialogueSystem
 {
-    public class DSC_Dialogue_ImageGroupController : BaseDialogueUIGroup
+    public class DSC_Dialogue_ImageGroupController : UIGroupController<DSC_Dialogue_ImageController>
     {
         #region Variable
 
@@ -12,14 +13,15 @@ namespace DSC.DialogueSystem
 #pragma warning disable 0649
 
         [SerializeField] protected DSC_Dialogue_DataController m_hDataController;
-        [SerializeField] protected DSC_Dialogue_ImageController[] m_arrImageController;
+        [SerializeField] protected List<DSC_Dialogue_ImageController> m_lstImageController;
 
 #pragma warning restore 0649
         #endregion
 
         #region Variable - Property
 
-        public override UIType GroupType { get { return UIType.Image; } }
+        public override UIType groupType { get { return UIType.Image; } }
+        protected override List<DSC_Dialogue_ImageController> uiControllerList { get { return m_lstImageController; } }
 
         #endregion
 
@@ -27,7 +29,7 @@ namespace DSC.DialogueSystem
 
         #region Base - Mono
 
-        private void Awake()
+        protected virtual void Awake()
         {
             if (m_hDataController && m_hDataController.DialogueEventDataList != null)
             {
@@ -38,7 +40,7 @@ namespace DSC.DialogueSystem
             }
         }
 
-        private void OnDestroy()
+        protected virtual void OnDestroy()
         {
             if (m_hDataController != null && m_hDataController.DialogueEventDataList != null)
             {
@@ -48,64 +50,39 @@ namespace DSC.DialogueSystem
 
         #endregion
 
-        #region Base - Override
-
-        public override void SetEnable(int nIndex, bool bEnable)
-        {
-            var hController = GetImageController(nIndex);
-            hController?.SetEnable(bEnable);
-        }
-
-        public override void SetAllEnable(bool bEnable)
-        {
-            if (!HasImageInArray())
-                return;
-
-            for (int i = 0; i < m_arrImageController.Length; i++)
-            {
-                var hImageController = m_arrImageController[i];
-                if (hImageController == null)
-                    continue;
-
-                hImageController.SetEnable(bEnable);
-            }
-        }
-
-        #endregion
-
         #region Main
 
-        public void SetSprite(int nIndex, Sprite hSprite)
+        public virtual void SetSprite(int nIndex, Sprite hSprite)
         {
             var hController = GetImageController(nIndex);
             hController?.SetImage(hSprite);
         }
 
-        public void SetPosition(int nIndex, Vector2 vPosition)
+        public virtual void SetPosition(int nIndex, Vector2 vPosition)
         {
             var hController = GetImageController(nIndex);
             hController?.SetPosition(vPosition);
         }
 
-        public void SetSize(int nIndex, Vector2 vSize)
+        public virtual void SetSize(int nIndex, Vector2 vSize)
         {
             var hController = GetImageController(nIndex);
             hController?.SetSize(vSize);
         }
 
-        public void SetSizeToNative(int nIndex)
+        public virtual void SetSizeToNative(int nIndex)
         {
             var hController = GetImageController(nIndex);
             hController?.SetSizeToNative();
         }
 
-        public void SetRotation(int nIndex,Vector3 vRotation)
+        public virtual void SetRotation(int nIndex,Vector3 vRotation)
         {
             var hController = GetImageController(nIndex);
             hController?.SetRotation(vRotation);
         }
 
-        public void SetColor(int nIndex,Color hColor)
+        public virtual void SetColor(int nIndex,Color hColor)
         {
             var hController = GetImageController(nIndex);
             hController?.SetColor(hColor);
@@ -115,19 +92,19 @@ namespace DSC.DialogueSystem
 
         #region Helper
 
-        protected bool HasImageInArray()
+        protected bool HasImageInList()
         {
-            return (m_arrImageController != null && m_arrImageController.Length > 0);
+            return (m_lstImageController != null && m_lstImageController.Count > 0);
         }
 
         protected DSC_Dialogue_ImageController GetImageController(int nIndex)
         {
             DSC_Dialogue_ImageController hResult = null;
 
-            if (nIndex < 0 || m_arrImageController == null ||  m_arrImageController.Length <= nIndex)
+            if (nIndex < 0 || m_lstImageController == null || m_lstImageController.Count <= nIndex)
                 return hResult;
 
-            hResult = m_arrImageController[nIndex];
+            hResult = m_lstImageController[nIndex];
             
             return hResult;
         }