+++ Scripts/Controllers/DSC_Dialogue_RawImageGroupController.cs
@@ -1,10 +1,11 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using DSC.UI;
 
 namespace DSC.DialogueSystem
 {
-    public class DSC_Dialogue_RawImageGroupController : BaseDialogueUIGroup
+    public class DSC_Dialogue_RawImageGroupController : UIGroupController<DSC_Dialogue_RawImageController>
     {
         #region Variable
 
@@ -12,14 +13,15 @@ namespace DSC.DialogueSystem
 #pragma warning disable 0649
 
         [SerializeField] protected DSC_Dialogue_DataController m_hDataController;
-        [SerializeField] protected DSC_Dialogue_RawImageController[] m_arrRawImageController;
+        [SerializeField] protected List<DSC_Dialogue_RawImageController> m_lstRawImageController;
 
 #pragma warning restore 0649
         #endregion
 
         #region Variable - Property
 
-        public override UIType GroupType { get { return UIType.RawImage; } }
+        public override UIType groupType { get { return UIType.RawImage; } }
+        protected override List<DSC_Dialogue_RawImageController> uiControllerList { get { return m_lstRawImageController; } }
 
         #endregion
 
@@ -27,7 +29,7 @@ namespace DSC.DialogueSystem
 
         #region Base - Mono
 
-        private void Awake()
+        protected virtual void Awake()
         {
             if (m_hDataController && m_hDataController.DialogueEventDataList != null)
             {
@@ -38,7 +40,7 @@ namespace DSC.DialogueSystem
             }
         }
 
-        private void OnDestroy()
+        protected virtual void OnDestroy()
         {
             if (m_hDataController != null && m_hDataController.DialogueEventDataList != null)
             {
@@ -48,60 +50,39 @@ namespace DSC.DialogueSystem
 
         #endregion
 
-        #region Base - Override
-
-        public override void SetEnable(int nIndex, bool bEnable)
-        {
-            var hController = GetRawImageController(nIndex);
-            hController?.SetEnable(bEnable);
-        }
-
-        public override void SetAllEnable(bool bEnable)
-        {
-            if (!HasRawImageInArray())
-                return;
-
-            for(int i = 0; i < m_arrRawImageController.Length; i++)
-            {
-                m_arrRawImageController[i]?.SetEnable(bEnable);
-            }
-        }
-
-        #endregion
-
         #region Main
 
-        public void SetTexture(int nIndex, Texture hTexture)
+        public virtual void SetTexture(int nIndex, Texture hTexture)
         {
             var hController = GetRawImageController(nIndex);
             hController?.SetImage(hTexture);
         }
 
-        public void SetPosition(int nIndex, Vector2 vPosition)
+        public virtual void SetPosition(int nIndex, Vector2 vPosition)
         {
             var hController = GetRawImageController(nIndex);
             hController?.SetPosition(vPosition);
         }
 
-        public void SetSize(int nIndex, Vector2 vSize)
+        public virtual void SetSize(int nIndex, Vector2 vSize)
         {
             var hController = GetRawImageController(nIndex);
             hController?.SetSize(vSize);
         }
 
-        public void SetSizeToNative(int nIndex)
+        public virtual void SetSizeToNative(int nIndex)
         {
             var hController = GetRawImageController(nIndex);
             hController?.SetSizeToNative();
         }
 
-        public void SetRotation(int nIndex, Vector3 vRotation)
+        public virtual void SetRotation(int nIndex, Vector3 vRotation)
         {
             var hController = GetRawImageController(nIndex);
             hController?.SetRotation(vRotation);
         }
 
-        public void SetColor(int nIndex, Color hColor)
+        public virtual void SetColor(int nIndex, Color hColor)
         {
             var hController = GetRawImageController(nIndex); ;
             hController?.SetColor(hColor);
@@ -111,19 +92,19 @@ namespace DSC.DialogueSystem
 
         #region Helper
 
-        protected bool HasRawImageInArray()
+        protected bool HasRawImageInList()
         {
-            return (m_arrRawImageController != null && m_arrRawImageController.Length > 0);
+            return (m_lstRawImageController != null && m_lstRawImageController.Count > 0);
         }
 
         protected DSC_Dialogue_RawImageController GetRawImageController(int nIndex)
         {
             DSC_Dialogue_RawImageController hResult = null;
 
-            if (nIndex < 0 || m_arrRawImageController == null || m_arrRawImageController.Length <= nIndex)
+            if (nIndex < 0 || m_lstRawImageController == null || m_lstRawImageController.Count <= nIndex)
                 return hResult;
 
-            hResult = m_arrRawImageController[nIndex];
+            hResult = m_lstRawImageController[nIndex];
 
             return hResult;
         }